refactor: module naming updated
