+++ jac-cloud/jac_cloud/plugin/implementation/websocket.py
@@ -15,7 +15,7 @@ from broadcaster import Broadcast
 
 from fastapi import APIRouter, WebSocket
 
-from jaclang.runtimelib.feature import JacMachine as Jac
+from jaclang.runtimelib.machine import JacMachine as Jac
 
 from orjson import dumps, loads
