+++ jac/jaclang/runtimelib/tests/fixtures/graph_purger.jac
@@ -1,4 +1,4 @@
-import:py from jaclang.runtimelib.feature {JacMachine as Jac}
+import:py from jaclang.runtimelib.machine {JacMachine as Jac}
 
 node A {
     has id: int;
