+++ jac/jaclang/tests/fixtures/dynamic_architype.jac
@@ -1,4 +1,4 @@
-import:py from jaclang.runtimelib.feature {JacMachine, JacMachineState}
+import:py from jaclang.runtimelib.machine {JacMachine, JacMachineState}
 import:jac from bar {Item}
 
 node test_node {
