+++ jac/jaclang/tests/test_language.py
@@ -10,7 +10,7 @@ from jaclang import JacMachine as Jac
 from jaclang.cli import cli
 from jaclang.compiler.passes.main.schedules import py_code_gen_typed
 from jaclang.compiler.program import JacProgram
-from jaclang.runtimelib.machine import JacMachineState
+from jaclang.runtimelib.machinestate import JacMachineState
 from jaclang.utils.test import TestCase
 
 import pytest
